Handle processFilter
