+++ src/PrestaShopBundle/Bridge/Helper/HelperListConfiguration.php
@@ -51,6 +51,11 @@ class HelperListConfiguration
     /**
      * @var string
      */
+    public $controllerNameLegacy;
+
+    /**
+     * @var string
+     */
     public $identifier;
 
     /**
