+++ src/PrestaShopBundle/Bridge/Smarty/SmartyTrait.php
@@ -1,0 +1,215 @@
+<?php
+
+namespace PrestaShopBundle\Bridge\Smarty;
+
+use \Media;
+use PrestaShopBundle\Bridge\Controller\ControllerConfiguration;
+use Symfony\Component\HttpFoundation\Response;
+use \Tools;
+
+trait SmartyTrait
+{
+    public function renderSmarty(string $content, ControllerConfiguration $controllerConfiguration): Response
+    {
+        $this->setMedia();
+
+        return $this->get('prestashop.core.bridge.smarty_bridge')->render($content, $controllerConfiguration);
+    }
+
+    public function setMedia($isNewTheme = false): void
+    {
+        $adminWebpath = '';
+
+        if (defined('_PS_ADMIN_DIR_')) {
+            $adminWebpath = preg_replace(
+                '/^' . preg_quote(DIRECTORY_SEPARATOR, '/') . '/', '',
+                str_ireplace(_PS_CORE_DIR_, '', _PS_ADMIN_DIR_)
+            );
+        }
+
+        if ($this->getContext()->language->is_rtl) {
+            $this->addJS(_PS_JS_DIR_ . 'rtl.js');
+            $this->addCSS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/css/' . $this->getContext()->language->iso_code . '.css');
+        }
+
+        if ($isNewTheme) {
+            $this->addCSS(__PS_BASE_URI__ . $adminWebpath . '/themes/new-theme/public/theme.css', 'all', 1);
+            $this->addJS(__PS_BASE_URI__ . $adminWebpath . '/themes/new-theme/public/main.bundle.js');
+
+            // the multistore dropdown should be called only once, and only if multistore is used
+            if ($this->get('prestashop.adapter.multistore_feature')->isUsed()) {
+                $this->addJs(__PS_BASE_URI__ . $adminWebpath . '/themes/new-theme/public/multistore_dropdown.bundle.js');
+            }
+            $this->addJqueryPlugin(['chosen', 'fancybox']);
+        } else {
+            //Bootstrap
+            $this->addCSS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/css/theme.css', 'all', 0);
+            $this->addCSS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/css/vendor/titatoggle-min.css', 'all', 0);
+            $this->addCSS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/public/theme.css', 'all', 0);
+
+            // add Jquery 3 and its migration script
+            $this->addJs(_PS_JS_DIR_ . 'jquery/jquery-3.5.1.min.js');
+            $this->addJs(_PS_JS_DIR_ . 'jquery/bo-migrate-mute.min.js');
+            $this->addJs(_PS_JS_DIR_ . 'jquery/jquery-migrate-3.1.0.min.js');
+            // implement $.browser object and live method, that has been removed since jquery 1.9
+            $this->addJs(_PS_JS_DIR_ . 'jquery/jquery.browser-0.1.0.min.js');
+            $this->addJs(_PS_JS_DIR_ . 'jquery/jquery.live-polyfill-1.1.2.min.js');
+
+            $this->addJqueryPlugin(['scrollTo', 'alerts', 'chosen', 'autosize', 'fancybox']);
+            $this->addJqueryPlugin('growl', null, false);
+            $this->addJqueryUI(['ui.slider', 'ui.datepicker']);
+
+            $this->addJS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/js/vendor/bootstrap.min.js');
+            $this->addJS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/js/vendor/modernizr.min.js');
+            $this->addJS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/js/modernizr-loads.js');
+            $this->addJS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/js/vendor/moment-with-langs.min.js');
+            $this->addJS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/public/bundle.js');
+
+            $this->addJS(_PS_JS_DIR_ . 'jquery/plugins/timepicker/jquery-ui-timepicker-addon.js');
+
+            if (!$this->controllerConfiguration->liteDisplay) {
+                $this->addJS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/js/help.js');
+            }
+
+            if (!Tools::getValue('submitFormAjax')) {
+                $this->addJS(_PS_JS_DIR_ . 'admin/notifications.js');
+            }
+
+            // Specific Admin Theme
+            $this->addCSS(__PS_BASE_URI__ . $adminWebpath . '/themes/' . self::DEFAULT_THEME . '/css/overrides.css', 'all', PHP_INT_MAX);
+        }
+
+        $this->addJS([
+            _PS_JS_DIR_ . 'admin.js?v=' . _PS_VERSION_, // TODO: SEE IF REMOVABLE
+            __PS_BASE_URI__ . $adminWebpath . '/themes/new-theme/public/cldr.bundle.js',
+            _PS_JS_DIR_ . 'tools.js?v=' . _PS_VERSION_,
+            __PS_BASE_URI__ . $adminWebpath . '/public/bundle.js',
+        ]);
+
+        Media::addJsDef([
+            'changeFormLanguageUrl' => $this->getContext()->link->getAdminLink(
+                'AdminEmployees',
+                true,
+                [],
+                ['action' => 'formLanguage']
+            ),
+        ]);
+        Media::addJsDef(['host_mode' => (defined('_PS_HOST_MODE_') && _PS_HOST_MODE_)]);
+        Media::addJsDef(['baseDir' => __PS_BASE_URI__]);
+        Media::addJsDef(['baseAdminDir' => __PS_BASE_URI__ . basename(_PS_ADMIN_DIR_) . '/']);
+        Media::addJsDef(['currency' => [
+            'iso_code' => $this->getContext()->currency->iso_code,
+            'sign' => $this->getContext()->currency->sign,
+            'name' => $this->getContext()->currency->name,
+            'format' => $this->getContext()->currency->format,
+        ]]);
+
+        Media::addJsDef([
+            'prestashop' => [
+                'debug' => _PS_MODE_DEV_,
+            ],
+        ]);
+
+        // Execute Hook AdminController SetMedia
+        $this->dispatchHook('actionAdminControllerSetMedia', []);
+    }
+
+    /**
+     * Adds a new stylesheet(s) to the page header.
+     */
+    public function addCSS($cssUri, $cssMediaType = 'all', $offset = null, $checkPath = true): void
+    {
+        if (!is_array($cssUri)) {
+            $cssUri = [$cssUri];
+        }
+
+        foreach ($cssUri as $cssFile => $media) {
+            if (is_string($cssFile) && strlen($cssFile) > 1) {
+                if ($checkPath) {
+                    $css_path = Media::getCSSPath($cssFile, $media);
+                } else {
+                    $css_path = [$cssFile => $media];
+                }
+            } else {
+                if ($checkPath) {
+                    $css_path = Media::getCSSPath($media, $cssMediaType);
+                } else {
+                    $css_path = [$media => $cssMediaType];
+                }
+            }
+
+            $key = is_array($css_path) ? key($css_path) : $css_path;
+            if ($css_path && (!isset($this->controllerConfiguration->cssFiles[$key]) || ($this->controllerConfiguration->cssFiles[$key] != reset($css_path)))) {
+                $size = count($this->controllerConfiguration->cssFiles);
+                if ($offset === null || $offset > $size || $offset < 0 || !is_numeric($offset)) {
+                    $offset = $size;
+                }
+
+                $this->controllerConfiguration->cssFiles = array_merge(array_slice($this->controllerConfiguration->cssFiles, 0, $offset), $css_path, array_slice($this->controllerConfiguration->cssFiles, $offset));
+            }
+        }
+    }
+
+    /**
+     * Adds a new JavaScript file(s) to the page header.
+     */
+    public function addJS($jsUri, $checkPath = true): void
+    {
+        if (!is_array($jsUri)) {
+            $jsUri = [$jsUri];
+        }
+
+        foreach ($jsUri as $jsFile) {
+            $jsFile = explode('?', $jsFile);
+            $version = '';
+            if (isset($jsFile[1]) && $jsFile[1]) {
+                $version = $jsFile[1];
+            }
+            $js_path = $jsFile = $jsFile[0];
+            if ($checkPath) {
+                $js_path = Media::getJSPath($jsFile);
+            }
+
+            if ($js_path && !in_array($js_path, $this->controllerConfiguration->jsFiles)) {
+                $this->controllerConfiguration->jsFiles[] = $js_path . ($version ? '?' . $version : '');
+            }
+        }
+    }
+
+    /**
+     * Adds jQuery plugin(s) to queued JS file list.
+     */
+    public function addJqueryPlugin($name, $folder = null, $css = true): void
+    {
+        if (!is_array($name)) {
+            $name = [$name];
+        }
+
+        foreach ($name as $plugin) {
+            $plugin_path = Media::getJqueryPluginPath($plugin, $folder);
+
+            if (!empty($plugin_path['js'])) {
+                $this->addJS($plugin_path['js'], false);
+            }
+            if ($css && !empty($plugin_path['css'])) {
+                $this->addCSS(key($plugin_path['css']), 'all', null, false);
+            }
+        }
+    }
+
+    /**
+     * Adds jQuery UI component(s) to queued JS file list.
+     */
+    public function addJqueryUI($component, $theme = 'base', $check_dependencies = true): void
+    {
+        if (!is_array($component)) {
+            $component = [$component];
+        }
+
+        foreach ($component as $ui) {
+            $ui_path = Media::getJqueryUIPath($ui, $theme, $check_dependencies);
+            $this->addCSS($ui_path['css'], 'all');
+            $this->addJS($ui_path['js'], false);
+        }
+    }
+}
